Move DatabaseResolver to separate file
